Add files via upload
From the pie chart, we can see that episodes 6, 5, and 4 are most preferred by the viewers having 17.49%, 34.61%, and 24.43% respectively of rating 1 as compared to other episodes.
So, I narrowed my focus to these 3 columns and dig deeper into it, I further calculated and plotted bar graph to see how many females and males have given rank 1 to the episode 6,5 and 4.
a. For episode 6, 77 females gave a rating 1, 10 more than males, hence that this episode was liked by females more.
b. For episode 5, 152 males gave rating1 and while 131 were females, this shows that this episode was liked by males much more than females.
c. For episode 4, 112 males gave rating 1 and while only 90 females gave it 1st rating, this shows that this episode was liked by males ~56% more than the females.
I further explored that which age group prefers which episode the most (rated 1) and plotted pie chart for visualization and to analyze.
a. 33.33% of people gave rating 1 for episode 6 and they are between the age group '30-44'. Meaning that 33.33% of people of age group 30-44 like episode 6 the most.
b. Whilst, 31.8% of people of age group 45-60 prefers episode 5 the most.
c. Whereas, 34.65% of people liked episode 4 the most who belong to the 45-60 age group.

Since episode 5 is liked the most by people, I went to a further level to find that people based in 'SOUTH ATLANTIC' have given the rating 1 the most for the episode 5. 
Hence people in this area like the 5th episode more than any other episode by the bar graph.
Hence, all these observations guide us to conclude that, episode 5 of Star Wars is most liked by the people. 
And that males like this episode more than the females and people who are of age group between 45-60 prefers the episode 5 the most as compared to any other episode.
Furthermore, ~19% of people from 'SOUTH ATLANTIC' like episode 5.
66% of the viewers who have seen any of 6 films are Star Wars fans. And only 284 people aren't a fan amongst the viewers.
The bar graph summarizes that a whopping 64% of people aren't familiar with the expanded universe amongst the fans of Star Wars, and thus only a minor percentage of 33.2-star war fans are familiar with it.
We can conclude that 310 out of 549 females prefer the character 'Yoda' 'very preferably', which is 56.5% of all females. On the other hand, 62.5% of males prefer the character 'Han Solo' 'very preferably'.
So, more than half the population of males and females prefer 'Han Solo' and 'Yoda' respectively.
125 people of age group 18-29,157 people of age between 30-44 prefers 'Yoda'. While 180 and 151 people of age between 45-60 and above 60 respectively 'very preferably' favors 'Han Solo'.
Thus, 'Han Solo' is the most preferred character when compared to all age groups.
‘Han Solo’ is favored ‘very preferably’ by associate degree and bachelor degree holders in large numbers as compared to any other character of the film.
Location and character have a relationship as people from East Northcentral, West Northcentral and South Atlantic likes 'Yoda' character 'very preferably' as compared to any other character.